fixed decimal to integer migration

+++ db/migrate/20220320025652_create_reviews.rb
@@ -1,7 +1,7 @@
 class CreateReviews < ActiveRecord::Migration[6.1]
   def change
     create_table :reviews do |t|
-      t.numeric :recommended
+      t.integer :recommended
       t.string :comment
       t.belongs_to :lesson, null: false, foreign_key: true
     end
